grc builds

+++ GNU_Radio/testing/trash_files/read_data.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 def bits_to_text(bits):
     chars = []
     for i in range(0, len(bits), 8):
@@ -9,7 +10,7 @@ def bits_to_text(bits):
     return ''.join(chars)
 
 # Load file
-raw_data = np.fromfile("bits_read_in.bin", dtype=np.complex64)
+raw_data = np.fromfile("bits_read_in.bin", dtype=np.uint8)
 print("Raw data loaded from file:\n", raw_data)
 
 # Conver raw data to bits
